pyproject.toml
[tool.poetry]
name = "books-api"
version = "0.1.0"
description = ""
authors = ["VladimirKhrabrov <khrabrov.v@brandsgo.digital>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
